+++ samples/hello/hello_slider.rb
@@ -32,10 +32,7 @@ class HelloSlider
   
   body {
     shell {
-      row_layout(:vertical) {
-        fill true
-        center true
-      }
+      fill_layout :vertical
       
       text 'Hello, Slider!'